Add of Penalty, Shield, Bonus

+++ AndroidGame/app/src/main/java/taing/tran/vivier/androidgame/Artefact/Artifact.java
@@ -1,7 +1,7 @@
 package taing.tran.vivier.androidgame.Artefact;
 
 public interface Artifact {
-    int damage();
+    double damage();
     int speed();
     int health();
 }
